fix arch observer bug
The ARCH observer had a small bug in which it exposes the final register
state of the previous trace, rather than the initial register state
of the current trace. This is due to the fact that the method `self.reset_model()`
is called before the input is loaded (`self._load_input(input_)`).

This patch just transposes the two to fix the bug.